change paths and location of existing files
