+++ App/city-feedback/src/main/java/com/example/city_feedback/authentication/application/services/UserService.java
@@ -1,8 +1,8 @@
-package com.example.city_feedback.services;
+package com.example.city_feedback.authentication.application.services;
 
 import com.example.city_feedback.authentication.application.dto.UserRegistrationDto;
 import com.example.city_feedback.authentication.exceptions.InvalidInputException;
-import com.example.city_feedback.domain.User;
+import com.example.city_feedback.authentication.domain.models.User;
 import org.springframework.security.core.userdetails.UserDetailsService;
 
 public interface UserService extends UserDetailsService {
