+++ App/city-feedback/src/main/java/com/example/city_feedback/authentication/domain/models/Citizen.java
@@ -1,0 +1,7 @@
+package com.example.city_feedback.authentication.domain.models;
+
+import com.example.city_feedback.complaintManagement.domain.models.Complaint;
+
+public class Citizen extends User {
+    private Complaint[] complaints;
+}
