+++ App/city-feedback/src/main/java/com/example/city_feedback/authentication/domain/models/CityEmployee.java
@@ -1,4 +1,4 @@
-package com.example.city_feedback.domain;
+package com.example.city_feedback.authentication.domain.models;
 
 public class CityEmployee extends User {
     private String role;
