+++ App/city-feedback/src/main/java/com/example/city_feedback/authentication/infrastructure/repositories/UserRepository.java
@@ -1,6 +1,6 @@
-package com.example.city_feedback.repositories;
+package com.example.city_feedback.authentication.infrastructure.repositories;
 
-import com.example.city_feedback.domain.User;
+import com.example.city_feedback.authentication.domain.models.User;
 import org.springframework.data.jpa.repository.JpaRepository;
 import org.springframework.stereotype.Repository;
