+++ App/city-feedback/src/main/java/com/example/city_feedback/complaintManagement/application/services/ComplaintService.java
@@ -1,0 +1,4 @@
+package com.example.city_feedback.complaintManagement.application.services;
+
+public class ComplaintService {
+}
