+++ App/city-feedback/src/main/java/com/example/city_feedback/complaintManagement/application/services/ComplaintValidator.java
@@ -1,4 +1,4 @@
-package com.example.city_feedback.services;
+package com.example.city_feedback.complaintManagement.application.services;
 
 import java.util.function.Predicate;
