+++ App/city-feedback/src/main/java/com/example/city_feedback/complaintManagement/domain/models/Category.java
@@ -1,4 +1,4 @@
-package com.example.city_feedback.domain;
+package com.example.city_feedback.complaintManagement.domain.models;
 
 public class Category {
     private int id;
