--- App/city-feedback/src/main/java/com/example/city_feedback/config/LocaleConfig.java
@@ -1,4 +1,0 @@
-package com.example.city_feedback.config;
-
-public class LocaleConfig {
-}
