--- App/city-feedback/src/main/java/com/example/city_feedback/domain/Citizen.java
@@ -1,5 +1,0 @@
-package com.example.city_feedback.domain;
-
-public class Citizen extends User {
-    private Complaint[] complaints;
-}
