+++ App/city-feedback/src/main/java/com/example/city_feedback/progressManagement/application/services/ComplaintProgressService.java
@@ -1,0 +1,4 @@
+package com.example.city_feedback.progressManagement.application.services;
+
+public class ComplaintProgressService {
+}
