+++ App/city-feedback/src/main/java/com/example/city_feedback/progressManagement/infrastructure/repositories/ProgressRepository.java
@@ -1,6 +1,6 @@
-package com.example.city_feedback.repositories;
+package com.example.city_feedback.progressManagement.infrastructure.repositories;
 
-public class ComplaintProgressRepository {
+public class ProgressRepository {
 
     public String findCurrentStatus(int complaintId) {
         return null;
