--- App/city-feedback/src/main/java/com/example/city_feedback/services/ComplaintProgressService.java
@@ -1,4 +1,0 @@
-package com.example.city_feedback.services;
-
-public class ComplaintProgressService {
-}
