+++ App/city-feedback/src/test/java/com/example/city_feedback/authentication/application/AuthenticationServiceTest.java
@@ -1,10 +1,10 @@
-package com.example.city_feedback;
+package com.example.city_feedback.authentication.application;
 
 import com.example.city_feedback.authentication.application.dto.UserRegistrationDto;
 import com.example.city_feedback.authentication.exceptions.InvalidInputException;
-import com.example.city_feedback.domain.User;
-import com.example.city_feedback.repositories.UserRepository;
-import com.example.city_feedback.services.AuthenticationService;
+import com.example.city_feedback.authentication.domain.models.User;
+import com.example.city_feedback.authentication.infrastructure.repositories.UserRepository;
+import com.example.city_feedback.authentication.application.services.AuthenticationService;
 import org.junit.jupiter.api.BeforeEach;
 import org.junit.jupiter.api.DisplayName;
 import org.junit.jupiter.api.Test;
